requirements.txt
aiofiles==23.1.0
aiosmtplib==2.0.2
aiosqlite==0.17.0
anyio==3.7.0
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Django==4.2.2
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.96.1
fastapi-mail==1.2.8
Flask==2.3.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
Pillow==9.5.0
pydantic==1.10.9
PyJWT==2.7.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tortoise-orm==0.19.3
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
